Refactor RedditClient to be a configured Spring bean
