+++ server/src/main/java/com/feeder/server/ApplicationProperties.java
@@ -2,9 +2,9 @@ package com.feeder.server;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.stereotype.Component;
 
-@Configuration
+@Component
 public class ApplicationProperties {
 
     private String redditUsername;
